+++ backend-idea/src/main/kotlin/io/github/amichne/kast/idea/backend/KastPluginBackendClose.kt
@@ -1,0 +1,14 @@
+package io.github.amichne.kast.idea.backend
+
+internal fun KastPluginBackend.closeResources() {
+    val failures = listOf(
+        runCatching(referenceContinuations::close).exceptionOrNull(),
+        runCatching(diagnosticContinuations::close).exceptionOrNull(),
+        runCatching(relationshipContinuations::close).exceptionOrNull(),
+        runCatching(workspaceFilePaging::close).exceptionOrNull(),
+    ).filterNotNull()
+    failures.firstOrNull()?.let { first ->
+        failures.drop(1).forEach(first::addSuppressed)
+        throw first
+    }
+}
