+++ backend-idea/src/test/kotlin/io/github/amichne/kast/idea/backend/NativeSemanticGraphBackendTest.kt
@@ -1,6 +1,5 @@
 package io.github.amichne.kast.idea
 
-import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.project.Project
 import com.intellij.testFramework.junit5.TestApplication
 import com.intellij.testFramework.junit5.fixture.TestFixture
@@ -12,32 +11,19 @@ import io.github.amichne.kast.api.contract.ServerLimits
 import io.github.amichne.kast.api.contract.query.SemanticGraphPath
 import io.github.amichne.kast.api.contract.query.SemanticGraphQuery
 import io.github.amichne.kast.api.contract.result.SemanticGraphRelationKind
-import io.github.amichne.kast.api.contract.result.SemanticGraphGeneration
 import io.github.amichne.kast.api.contract.result.SemanticGraphSourcePath
 import io.github.amichne.kast.api.contract.result.SemanticGraphSymbolKind
 import io.github.amichne.kast.api.contract.result.SemanticGraphVisibility
-import io.github.amichne.kast.api.protocol.ConflictException
 import io.github.amichne.kast.api.protocol.ValidationException
 import io.github.amichne.kast.api.validation.parsed
 import io.github.amichne.kast.idea.backend.KastPluginBackend
 import io.github.amichne.kast.indexstore.store.SqliteSourceIndexStore
-import kotlinx.coroutines.CancellationException
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.async
 import kotlinx.coroutines.runBlocking
-import kotlinx.coroutines.supervisorScope
-import kotlinx.coroutines.withTimeout
 import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.io.TempDir
 import java.nio.file.Path
-import java.util.concurrent.CountDownLatch
-import java.util.concurrent.TimeUnit
-import java.util.concurrent.atomic.AtomicBoolean
-import java.util.concurrent.atomic.AtomicInteger
-import java.util.concurrent.atomic.AtomicLong
 
 @TestApplication
 class NativeSemanticGraphBackendTest {
@@ -68,18 +54,6 @@ class NativeSemanticGraphBackendTest {
             }
 
             fun construct(): Constructed = Constructed(1)
-        """
-
-        private const val boundarySource = """
-            package demo
-
-            fun reachBoundary(): BoundaryTarget = BoundaryTarget()
-        """
-
-        private const val boundaryTarget = """
-            package demo
-
-            class BoundaryTarget
         """
 
         private const val leftType = """
@@ -141,8 +115,6 @@ class NativeSemanticGraphBackendTest {
     private val moduleFixture = projectFixture.moduleFixture("main")
     private val sourceRootFixture = moduleFixture.sourceRootFixture()
     private val canonicalFileFixture = sourceRootFixture.psiFileFixture("Canonical.kt", canonicalSource)
-    private val boundarySourceFixture = sourceRootFixture.psiFileFixture("BoundarySource.kt", boundarySource)
-    private val boundaryTargetFixture = sourceRootFixture.psiFileFixture("BoundaryTarget.kt", boundaryTarget)
     private val leftTypeFixture = sourceRootFixture.psiFileFixture("LeftType.kt", leftType)
     private val rightTypeFixture = sourceRootFixture.psiFileFixture("RightType.kt", rightType)
     private val leftTypeConsumerFixture = sourceRootFixture.psiFileFixture("LeftTypeConsumer.kt", leftTypeConsumer)
@@ -208,324 +180,6 @@ class NativeSemanticGraphBackendTest {
                         relation.resolvedTargetKey in constructorKeys
                 },
             )
-        }
-    }
-
-    @Test
-    fun `source scope excludes unselected targets and widens additively`() = runBlocking {
-        val project = projectFixture.get()
-        val sourceFile = boundarySourceFixture.get()
-        val targetFile = boundaryTargetFixture.get()
-        waitUntilIndexesAreReady(project)
-        val workspaceRoot = Path.of(sourceFile.virtualFile.path).toRealPath().parent
-        val sourcePath = SemanticGraphPath.parse(sourceFile.virtualFile.path)
-        val targetPath = SemanticGraphPath.parse(targetFile.virtualFile.path)
-        fun query(path: SemanticGraphPath) = SemanticGraphQuery(filePaths = listOf(path)).parsed()
-        SqliteSourceIndexStore(storeRoot).use { store ->
-            store.ensureSchema()
-            KastPluginBackend(
-                project = project,
-                workspaceRoot = workspaceRoot,
-                limits = limits(),
-                semanticGraphStore = store,
-                psiGeneration = { 1L },
-            ).use { backend ->
-                val result = backend.semanticGraph(query(sourcePath))
-                assertTrue(result.coverage.omittedExternalTargetCount.value > 0)
-                val excluded = store.readSemanticGraph(listOf(SemanticGraphSourcePath.parse("BoundarySource.kt")))
-                assertTrue(excluded.boundarySymbols.none { symbol -> symbol.name.value == "BoundaryTarget" })
-                assertTrue(excluded.relations.none { relation -> relation.targetKey.value.contains("BoundaryTarget") })
-                backend.semanticGraph(query(targetPath))
-                backend.semanticGraph(query(sourcePath))
-            }
-            val snapshot = store.readSemanticGraph(listOf(SemanticGraphSourcePath.parse("BoundarySource.kt")))
-            val boundary = snapshot.boundarySymbols.single { symbol -> symbol.name.value == "BoundaryTarget" }
-            assertTrue(snapshot.relations.any { relation -> relation.targetKey == boundary.canonicalKey })
-        }
-    }
-
-    @Test
-    fun `multi file refresh releases IDEA read access between files`() = runBlocking {
-        val project = projectFixture.get()
-        val files = listOf(canonicalFileFixture.get(), boundaryTargetFixture.get())
-        waitUntilIndexesAreReady(project)
-        val workspaceRoot = Path.of(files.first().virtualFile.path).toRealPath().parent
-        val firstReadEntered = CountDownLatch(1)
-        val releaseFirstRead = CountDownLatch(1)
-        val writeStarted = CountDownLatch(1)
-        val writeCompleted = CountDownLatch(1)
-        val readCount = AtomicInteger()
-        val psiGeneration = AtomicLong(1)
-        val secondReadObservedCompletedWrite = AtomicBoolean(false)
-        val observer = IdeaReadEpochObserver { kind ->
-            if (kind != IdeaReadEpochKind.SEMANTIC_GRAPH) return@IdeaReadEpochObserver
-            when (readCount.incrementAndGet()) {
-                1 -> {
-                    firstReadEntered.countDown()
-                    assertTrue(releaseFirstRead.await(10, TimeUnit.SECONDS))
-                }
-                2 -> secondReadObservedCompletedWrite.set(writeCompleted.count == 0L)
-            }
-        }
-
-        SqliteSourceIndexStore(storeRoot).use { store ->
-            store.ensureSchema()
-            KastPluginBackend(
-                project = project,
-                workspaceRoot = workspaceRoot,
-                limits = limits(),
-                semanticGraphStore = store,
-                psiGeneration = psiGeneration::get,
-                readEpochObserver = observer,
-            ).use { backend ->
-                supervisorScope {
-                    val refresh = async(Dispatchers.Default) {
-                        backend.semanticGraph(
-                            SemanticGraphQuery(
-                                filePaths = files.map { file -> SemanticGraphPath.parse(file.virtualFile.path) },
-                            ).parsed(),
-                        )
-                    }
-                    assertTrue(firstReadEntered.await(10, TimeUnit.SECONDS))
-
-                    val application = ApplicationManager.getApplication()
-                    application.invokeLater {
-                        writeStarted.countDown()
-                        application.runWriteAction {
-                            psiGeneration.incrementAndGet()
-                            writeCompleted.countDown()
-                        }
-                    }
-                    assertTrue(writeStarted.await(10, TimeUnit.SECONDS))
-                    assertFalse(writeCompleted.await(100, TimeUnit.MILLISECONDS))
-
-                    releaseFirstRead.countDown()
-                    val failure = runCatching { refresh.await() }.exceptionOrNull()
-                    assertTrue(writeCompleted.await(10, TimeUnit.SECONDS))
-                    assertTrue(failure is ConflictException, "expected PSI generation conflict, got $failure")
-                    assertTrue(store.semanticGraphSourcePaths().isEmpty())
-                }
-            }
-        }
-
-        assertEquals(files.size, readCount.get())
-        assertTrue(
-            secondReadObservedCompletedWrite.get(),
-            "A pending IDEA write action must complete before the next semantic graph file read",
-        )
-    }
-
-    @Test
-    fun `cancelling multi file refresh leaves the committed graph unchanged`() = runBlocking {
-        val project = projectFixture.get()
-        val files = listOf(canonicalFileFixture.get(), boundaryTargetFixture.get())
-        waitUntilIndexesAreReady(project)
-        val workspaceRoot = Path.of(files.first().virtualFile.path).toRealPath().parent
-        val firstReadEntered = CountDownLatch(1)
-        val releaseFirstRead = CountDownLatch(1)
-        val readCount = AtomicInteger()
-        val observer = IdeaReadEpochObserver { kind ->
-            if (kind == IdeaReadEpochKind.SEMANTIC_GRAPH && readCount.incrementAndGet() == 1) {
-                firstReadEntered.countDown()
-                assertTrue(releaseFirstRead.await(10, TimeUnit.SECONDS))
-            }
-        }
-
-        SqliteSourceIndexStore(storeRoot).use { store ->
-            store.ensureSchema()
-            KastPluginBackend(
-                project = project,
-                workspaceRoot = workspaceRoot,
-                limits = limits(),
-                semanticGraphStore = store,
-                psiGeneration = { 1L },
-                readEpochObserver = observer,
-            ).use { backend ->
-                val refresh = async(Dispatchers.Default) {
-                    backend.semanticGraph(
-                        SemanticGraphQuery(
-                            filePaths = files.map { file -> SemanticGraphPath.parse(file.virtualFile.path) },
-                        ).parsed(),
-                    )
-                }
-                assertTrue(firstReadEntered.await(10, TimeUnit.SECONDS))
-
-                refresh.cancel()
-                releaseFirstRead.countDown()
-                val failure = runCatching { refresh.await() }.exceptionOrNull()
-
-                assertTrue(failure is CancellationException, "expected cancellation, got $failure")
-                assertEquals(1, readCount.get())
-                assertTrue(store.semanticGraphSourcePaths().isEmpty())
-            }
-        }
-    }
-
-    @Test
-    fun `concurrent removal prevents an older refresh from resurrecting cached nodes`() = runBlocking {
-        val project = projectFixture.get()
-        val sourceFile = canonicalFileFixture.get()
-        waitUntilIndexesAreReady(project)
-        val workspaceRoot = Path.of(sourceFile.virtualFile.path).toRealPath().parent
-        val sourcePath = SemanticGraphPath.parse(sourceFile.virtualFile.path)
-
-        SqliteSourceIndexStore(storeRoot).use { store ->
-            store.ensureSchema()
-            KastPluginBackend(
-                project = project,
-                workspaceRoot = workspaceRoot,
-                limits = limits(),
-                semanticGraphStore = store,
-                psiGeneration = { 1L },
-            ).use { backend ->
-                val seeded = backend.semanticGraph(SemanticGraphQuery(filePaths = listOf(sourcePath)).parsed())
-                assertTrue(seeded.symbolCount.value > 0)
-            }
-            val seededGeneration = store.readGeneration()
-
-            val refreshReadEntered = CountDownLatch(1)
-            val releaseRefreshRead = CountDownLatch(1)
-            val observer = IdeaReadEpochObserver { kind ->
-                if (kind == IdeaReadEpochKind.SEMANTIC_GRAPH) {
-                    refreshReadEntered.countDown()
-                    assertTrue(releaseRefreshRead.await(10, TimeUnit.SECONDS))
-                }
-            }
-            KastPluginBackend(
-                project = project,
-                workspaceRoot = workspaceRoot,
-                limits = limits(),
-                semanticGraphStore = store,
-                psiGeneration = { 1L },
-                readEpochObserver = observer,
-            ).use { backend ->
-                supervisorScope {
-                    val refresh = async(Dispatchers.Default) {
-                        backend.semanticGraph(
-                            SemanticGraphQuery(
-                                filePaths = listOf(sourcePath),
-                                expectedGeneration = SemanticGraphGeneration(seededGeneration.value),
-                            ).parsed(),
-                        )
-                    }
-                    assertTrue(refreshReadEntered.await(10, TimeUnit.SECONDS))
-                    val removal = async(Dispatchers.Default) {
-                        backend.semanticGraph(
-                            SemanticGraphQuery(
-                                filePaths = emptyList(),
-                                removedFilePaths = listOf(sourcePath),
-                                expectedGeneration = SemanticGraphGeneration(seededGeneration.value),
-                            ).parsed(),
-                        )
-                    }
-
-                    withTimeout(10_000) { removal.await() }
-                    releaseRefreshRead.countDown()
-                    val refreshFailure = runCatching { refresh.await() }.exceptionOrNull()
-
-                    assertTrue(refreshFailure is ConflictException, "expected generation conflict, got $refreshFailure")
-                    assertTrue(refreshFailure?.message.orEmpty().contains("retry"))
-                    assertTrue(store.semanticGraphSourcePaths().isEmpty())
-                }
-            }
-        }
-    }
-
-    @Test
-    fun `removal only refresh rejects a changed PSI generation without deleting cached nodes`() = runBlocking {
-        val project = projectFixture.get()
-        val sourceFile = canonicalFileFixture.get()
-        waitUntilIndexesAreReady(project)
-        val workspaceRoot = Path.of(sourceFile.virtualFile.path).toRealPath().parent
-        val sourcePath = SemanticGraphPath.parse(sourceFile.virtualFile.path)
-
-        SqliteSourceIndexStore(storeRoot).use { store ->
-            store.ensureSchema()
-            KastPluginBackend(
-                project = project,
-                workspaceRoot = workspaceRoot,
-                limits = limits(),
-                semanticGraphStore = store,
-                psiGeneration = { 1L },
-            ).use { backend ->
-                backend.semanticGraph(SemanticGraphQuery(filePaths = listOf(sourcePath)).parsed())
-            }
-            val generation = store.readGeneration()
-            val psiGeneration = AtomicLong()
-
-            val failure = KastPluginBackend(
-                project = project,
-                workspaceRoot = workspaceRoot,
-                limits = limits(),
-                semanticGraphStore = store,
-                psiGeneration = psiGeneration::incrementAndGet,
-            ).use { backend ->
-                runCatching {
-                    backend.semanticGraph(
-                        SemanticGraphQuery(
-                            filePaths = emptyList(),
-                            removedFilePaths = listOf(sourcePath),
-                            expectedGeneration = SemanticGraphGeneration(generation.value),
-                        ).parsed(),
-                    )
-                }.exceptionOrNull()
-            }
-
-            assertTrue(failure is ConflictException, "expected PSI generation conflict, got $failure")
-            assertEquals(generation, store.readGeneration())
-            assertEquals(setOf(SemanticGraphSourcePath.parse(sourceFile.name)), store.semanticGraphSourcePaths())
-        }
-    }
-
-    @Test
-    fun `stale expected generation rejects refresh before IDEA extraction`() = runBlocking {
-        val project = projectFixture.get()
-        val sourceFile = canonicalFileFixture.get()
-        waitUntilIndexesAreReady(project)
-        val workspaceRoot = Path.of(sourceFile.virtualFile.path).toRealPath().parent
-        val sourcePath = SemanticGraphPath.parse(sourceFile.virtualFile.path)
-
-        SqliteSourceIndexStore(storeRoot).use { store ->
-            store.ensureSchema()
-            val seededGeneration = KastPluginBackend(
-                project = project,
-                workspaceRoot = workspaceRoot,
-                limits = limits(),
-                semanticGraphStore = store,
-                psiGeneration = { 1L },
-            ).use { backend ->
-                backend.semanticGraph(SemanticGraphQuery(filePaths = listOf(sourcePath)).parsed()).generation
-            }
-            store.replaceSemanticGraphFiles(
-                updates = emptyList(),
-                removedPaths = listOf(SemanticGraphSourcePath.parse(sourceFile.name)),
-            )
-            val readCount = AtomicInteger()
-            val observer = IdeaReadEpochObserver { kind ->
-                if (kind == IdeaReadEpochKind.SEMANTIC_GRAPH) readCount.incrementAndGet()
-            }
-
-            val failure = KastPluginBackend(
-                project = project,
-                workspaceRoot = workspaceRoot,
-                limits = limits(),
-                semanticGraphStore = store,
-                psiGeneration = { 1L },
-                readEpochObserver = observer,
-            ).use { backend ->
-                runCatching {
-                    backend.semanticGraph(
-                        SemanticGraphQuery(
-                            filePaths = listOf(sourcePath),
-                            expectedGeneration = seededGeneration,
-                        ).parsed(),
-                    )
-                }.exceptionOrNull()
-            }
-
-            assertTrue(failure is ConflictException, "expected generation conflict, got $failure")
-            assertEquals(0, readCount.get())
-            assertTrue(store.semanticGraphSourcePaths().isEmpty())
         }
     }
