+++ backend-shared/src/main/kotlin/io/github/amichne/kast/shared/hierarchy/discovery/EdgeDiscoveryBudget.kt
