+++ cli-rs/src/agent/navigation/native_graph.rs
@@ -3,9 +3,9 @@ include!("native_graph/source_scope.rs");
 include!("native_graph/query.rs");
 include!("native_graph/database.rs");
 include!("native_graph/neighbors.rs");
-include!("native_graph/base_graph.rs");
-include!("native_graph/overlay_graph.rs");
+include!("native_graph/graph/base.rs");
+include!("native_graph/graph/overlay.rs");
 include!("native_graph/graph_algorithms.rs");
 include!("native_graph/partitions.rs");
 include!("native_graph/measurements.rs");
-include!("native_graph/tests.rs");
+include!("native_graph/tests/mod.rs");
