+++ cli-rs/src/agent/navigation/native_graph/graph/base.rs
