+++ cli-rs/src/agent/navigation/native_graph/graph/overlay.rs
